requirements.txt
setuptools
pywebview
pyinstaller; sys_platform != 'darwin'
py2app; sys_platform == 'darwin'
pyqt5 
pyqtwebengine
pywebview[qt]